Implement our own barrier as threads could die during migration.
If your application has the problem of dying threads during the
migration, you can use config.track_threads. If you want to use this
feature, and let the wf-userland library track the number of threads,
you only have to set the option and insert wf_thread_{birth,death}()
callbacks at the beginning and the end of your threads-entry functions.